Made all components that can be bootstrap bootstrap
