--- apps/frontend/src/routes/errorPage.css
@@ -1,7 +1,0 @@
-#error-page {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-}
