+++ apps/frontend/src/routes/highScore.css
@@ -1,7 +1,3 @@
-#root {
-  text-align: center;
-}
-
 .logo {
   height: 6em;
   padding: 1.5em;
