+++ apps/frontend/src/scss/custom.scss
@@ -26,3 +26,6 @@
 @import "bootstrap/scss/utilities/api";
 
 // 8. Add additional custom code here
+
+// Import any remaining bootstrap defaults
+@import "bootstrap/scss/bootstrap";
